Fix textarea width bug
The existing style block setting left the width
calculation up to the browser defaults and the
`cols` attribute. This can cause textareas to
break out of their container.

This restricts their width to that of the
container while allowing it to be dictated by the
number of columns needed up until then.